feat(messages): add MessagesPage and integrate messaging functionality in the app
